Created generic UI folder, and moved AdvancedScrollView there
