+++ Assets/_Application/_Twin/UI/AdvancedScrollView.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
 
-namespace Netherlands3D.Twin
+namespace Netherlands3D.Twin.UI
 {
     public class AdvancedScrollView : MonoBehaviour
     {
